requirements.txt
gunicorn
flask
google-generativeai